remove v1 URL builder and use URL builder v2
