+++ Kentico.Kontent.Management/ManagementClient.cs
@@ -25,7 +25,6 @@ namespace Kentico.Kontent.Management
 
         private readonly ActionInvoker _actionInvoker;
         private readonly EndpointUrlBuilder _urlBuilder;
-        private readonly EndpointUrlBuilderV2 _urlBuilderV2;
         private readonly IModelProvider _modelProvider;
 
         /// <summary>
@@ -55,17 +54,15 @@ namespace Kentico.Kontent.Management
             }
 
             _urlBuilder = new EndpointUrlBuilder(ManagementOptions);
-            _urlBuilderV2 = new EndpointUrlBuilderV2(ManagementOptions);
             _actionInvoker = new ActionInvoker(
                 new ManagementHttpClient(new DefaultResiliencePolicyProvider(ManagementOptions.MaxRetryAttempts), ManagementOptions.EnableResilienceLogic),
                 new MessageCreator(ManagementOptions.ApiKey));
             _modelProvider = ManagementOptions.ModelProvider ?? new ModelProvider();
         }
 
-        internal ManagementClient(EndpointUrlBuilder urlBuilder, EndpointUrlBuilderV2 urlBuilderV2, ActionInvoker actionInvoker, IModelProvider modelProvider = null)
+        internal ManagementClient(EndpointUrlBuilder urlBuilder, ActionInvoker actionInvoker, IModelProvider modelProvider = null)
         {
             _urlBuilder = urlBuilder ?? throw new ArgumentNullException(nameof(urlBuilder));
-            _urlBuilderV2 = urlBuilderV2 ?? throw new ArgumentNullException(nameof(urlBuilderV2));
             _actionInvoker = actionInvoker ?? throw new ArgumentNullException(nameof(actionInvoker));
             _modelProvider = modelProvider ?? new ModelProvider();
         }
@@ -353,7 +350,7 @@ namespace Kentico.Kontent.Management
         /// <returns>The <see cref="ListingResponseModel{AssetModel}"/> instance that represents the listing of assets.</returns>
         public async Task<ListingResponseModel<AssetModel>> ListAssetsAsync()
         {
-            var endpointUrl = _urlBuilderV2.BuildAssetsUrl();
+            var endpointUrl = _urlBuilder.BuildAssetsUrl();
             var response = await _actionInvoker.InvokeReadOnlyMethodAsync<AssetListingResponseServerModel>(endpointUrl, HttpMethod.Get);
 
             return new ListingResponseModel<AssetModel>(GetNextAssetListingPageAsync, response.Pagination?.Token, response.Assets);
@@ -361,7 +358,7 @@ namespace Kentico.Kontent.Management
 
         private async Task<IListingResponse<AssetModel>> GetNextAssetListingPageAsync(string continuationToken)
         {
-            var endpointUrl = _urlBuilderV2.BuildAssetsUrl();
+            var endpointUrl = _urlBuilder.BuildAssetsUrl();
             var headers = new Dictionary<string, string>();
             headers.Add("x-continuation", continuationToken);
             var response = await _actionInvoker.InvokeReadOnlyMethodAsync<AssetListingResponseServerModel>(endpointUrl, HttpMethod.Get, headers);
@@ -381,7 +378,7 @@ namespace Kentico.Kontent.Management
                 throw new ArgumentNullException(nameof(identifier));
             }
 
-            var endpointUrl = _urlBuilderV2.BuildAssetsUrl(identifier);
+            var endpointUrl = _urlBuilder.BuildAssetsUrl(identifier);
             var response = await _actionInvoker.InvokeReadOnlyMethodAsync<AssetModel>(endpointUrl, HttpMethod.Get);
 
             return response;
@@ -393,7 +390,7 @@ namespace Kentico.Kontent.Management
         /// <returns>Returns the hierarchy of asset folders beginning with the root level</returns>
         public async Task<AssetFolderList> GetAssetFoldersAsync()
         {
-            var endpointUrl = _urlBuilderV2.BuildAssetFoldersUrl();
+            var endpointUrl = _urlBuilder.BuildAssetFoldersUrl();
             var response = await _actionInvoker.InvokeReadOnlyMethodAsync<AssetFolderList>(endpointUrl, HttpMethod.Get);
 
             return response;
@@ -417,7 +414,7 @@ namespace Kentico.Kontent.Management
                 throw new ArgumentNullException(nameof(asset));
             }
 
-            var endpointUrl = _urlBuilderV2.BuildAssetsUrl(identifier);
+            var endpointUrl = _urlBuilder.BuildAssetsUrl(identifier);
             var response = await _actionInvoker.InvokeMethodAsync<AssetUpdateModel, AssetModel>(endpointUrl, HttpMethod.Put, asset);
 
             return response;
@@ -450,7 +447,7 @@ namespace Kentico.Kontent.Management
                 throw new ArgumentNullException(nameof(asset));
             }
 
-            var endpointUrl = _urlBuilderV2.BuildAssetsUrl();
+            var endpointUrl = _urlBuilder.BuildAssetsUrl();
             var response = await _actionInvoker.InvokeMethodAsync<AssetUpsertModel, AssetModel>(endpointUrl, HttpMethod.Post, asset);
 
             return response;
@@ -474,7 +471,7 @@ namespace Kentico.Kontent.Management
                 throw new ArgumentNullException(nameof(asset));
             }
 
-            var endpointUrl = _urlBuilderV2.BuildAssetsUrlFromExternalId(externalId);
+            var endpointUrl = _urlBuilder.BuildAssetsUrlFromExternalId(externalId);
             var response = await _actionInvoker.InvokeMethodAsync<AssetUpsertModel, AssetModel>(
                 endpointUrl,
                 HttpMethod.Put,
